feat: add core logic and shopping list endpoint

+++ backend/app/models.py
@@ -9,8 +9,8 @@ class Product(Base):
     name = Column(String, nullable=False)
     category = Column(String, nullable=True)
     unit = Column(String, nullable=True)
-    current_stock = Column(Float, default=0)
-    minimum_stock = Column(Float, default=1)
+    current_stock = Column(Integer, default=0)
+    minimum_stock = Column(Integer, default=1)
     created_at = Column(DateTime(timezone=True), server_default=func.now())
     updated_at = Column(DateTime(timezone=True), onupdate=func.now())
